+++ src/rendering/shaders/structs.inc.glsl
@@ -3,6 +3,7 @@
 
 #extension GL_EXT_buffer_reference : require
 #extension GL_EXT_scalar_block_layout : require
+#extension GL_EXT_shader_explicit_arithmetic_types_int16 : require
 
 struct VertexInput
 {
@@ -18,7 +19,7 @@ layout(buffer_reference, scalar) readonly buffer VertexBuffer
 
 layout(buffer_reference, scalar) readonly buffer IndexBuffer
 {
-    uint indices[];
+    uint16_t indices[];
 };
 
 struct Material {
